# docs/lead-import-api.openapi.yaml
openapi: 3.1.0
info:
  title: Lead Center – Lead API
  version: 1.1.0
  description: |
    Externe HTTP-Schnittstelle, über die ein anderes System (z. B. Claude Cowork) Leads ins
    Lead-Center **importiert**, **ausliest** und **aktualisiert**.

    Import (`POST /api/leads/import`): jeder Aufruf erzeugt einen Import-Batch
    (`import_logs`-Eintrag, `import_type: "api"`); alle angelegten Leads teilen dessen
    `source_import_id` und erscheinen mit Status `"imported"` ("neue Leads") als eine Gruppe in
    der Import-Historie (`/import`). Duplikate werden **strikt übersprungen**: existiert bereits
    ein Lead mit gleicher Domain, E-Mail, Telefonnummer oder (fuzzy) Firmenname, wird er nicht
    verändert — nur komplett neue Firmen werden angelegt.

    Lesen (`GET /api/leads`, `GET /api/leads/{id}`) und Aktualisieren (`PATCH /api/leads/{id}`)
    teilen sich denselben Bearer-Token. `PATCH` ändert nur whitelistete Stammdaten-Felder
    (kein status/assigned_to/deleted_at).

servers:
  - url: https://your-domain.example
    description: Produktion (Domain anpassen)
  - url: http://localhost:3000
    description: Lokale Entwicklung

security:
  - bearerAuth: []

paths:
  /api/leads/import:
    post:
      operationId: importLeads
      summary: Neue Leads anlegen (Batch)
      description: |
        Legt einen oder mehrere neue Leads an. Authentifizierung per Bearer-Token
        (`LEADS_IMPORT_API_KEY`). Pro Aufruf max. 10.000 Leads; praktischer Richtwert wegen des
        Request-Body-Limits (~4,5 MB) ≤ 1.000 Leads.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
            examples:
              einzelnerLead:
                summary: Ein Lead mit Ansprechpartner
                value:
                  source: claude-cowork
                  leads:
                    - company_name: Beispiel GmbH
                      website: beispiel.de
                      phone: 0571 123456
                      email: info@beispiel.de
                      city: Espelkamp
                      zip: "32339"
                      industry: Maschinenbau
                      contacts:
                        - name: Max Mustermann
                          role: Geschäftsführer
      responses:
        '201':
          description: Import verarbeitet (auch wenn 0 Leads neu sind, z. B. nur Duplikate).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
              examples:
                erfolg:
                  value:
                    success: true
                    import_log_id: fe3e1455-8158-41dc-a47f-edb248a21996
                    imported: 1
                    skipped: 0
                    duplicates: 0
                    updated: 0
                    archived: 0
                    errors: 0
                    contacts_imported: 1
                    error_details: []
        '400':
          description: |
            Ungültige Anfrage — kein/kaputtes JSON, `leads` fehlt/leer/kein Array,
            oder mehr als 10.000 Leads.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                leeresArray:
                  value:
                    error: "Feld 'leads' muss ein nicht-leeres Array sein."
                kaputtesJson:
                  value:
                    error: Ungueltiges JSON
        '401':
          description: Bearer-Token fehlt oder ist falsch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unauthorized:
                  value:
                    error: Unauthorized
        '500':
          description: Interner Fehler (z. B. Import-Log konnte nicht angelegt werden).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/leads:
    get:
      operationId: listLeads
      summary: Leads auflisten (filterbar, paginiert)
      description: |
        Listet Leads ohne gelöschte, neueste zuerst. Für die neuen Leads: `status=imported`.
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
          description: 'Status-Filter, z. B. `imported` (neue Leads), `qualified`, `customer`.'
          example: imported
        - name: vertical
          in: query
          schema: { type: string }
          description: Sparte/Filter.
        - name: q
          in: query
          schema: { type: string }
          description: Suche über Firmenname / Website / Ort / E-Mail / Telefon.
        - name: include
          in: query
          schema: { type: string }
          description: 'Kommasepariert: `contacts` und/oder `links` → reichert jeden Lead an.'
          example: contacts,links
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Liste passender Leads.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadListResponse'
        '401':
          description: Bearer-Token fehlt oder ist falsch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Interner Fehler / DB-Fehler.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/leads/{id}:
    get:
      operationId: getLead
      summary: Einzelnen Lead lesen (inkl. Kontakte + Links)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lead inkl. Kontakte + Links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadDetailResponse'
        '401':
          description: Bearer-Token fehlt oder ist falsch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Kein (nicht gelöschter) Lead mit dieser ID.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    patch:
      operationId: updateLead
      summary: Stammdaten eines Leads aktualisieren
      description: |
        Aktualisiert nur whitelistete Stammdaten-Felder (siehe `LeadUpdate`). Andere Keys
        (status, assigned_to, deleted_at, crm_status_id …) werden ignoriert. Telefon-Änderung
        setzt `phone_source="manual"`, Adress-Änderung setzt die Geokoordinaten zurück.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadUpdate'
            examples:
              telefonUndBranche:
                value:
                  phone: 0571 999999
                  industry: Maschinenbau
      responses:
        '200':
          description: Aktualisiert (auch bei No-Op).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '400':
          description: Kaputtes JSON oder Body ist kein Objekt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Bearer-Token fehlt oder ist falsch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Lead nicht gefunden.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Wert der Umgebungsvariable `LEADS_IMPORT_API_KEY`.

  schemas:
    ImportRequest:
      type: object
      required: [leads]
      properties:
        source:
          type: string
          description: |
            Optionale Quell-Bezeichnung. Erscheint als Batch-Name in der Import-Historie
            (`file_name = "api:<source>"`). Default: `claude-cowork`.
          example: claude-cowork
        leads:
          type: array
          minItems: 1
          maxItems: 10000
          description: Liste der anzulegenden Leads. Mehr als 10.000 → 400.
          items:
            $ref: '#/components/schemas/LeadInput'

    LeadInput:
      type: object
      required: [company_name]
      description: |
        Pflicht ist nur `company_name`. Felder werden serverseitig normalisiert
        (Telefon → `+49…`, E-Mail → lowercase, `website` → nackte Domain, `country` Default
        „Deutschland"). Leere Strings gelten als nicht gesetzt. Leads ohne `company_name`
        werden abgelehnt und in `error_details` aufgeführt.
      properties:
        company_name:
          type: string
          description: Firmenname (Pflichtfeld).
          example: Beispiel GmbH
        website:
          type: string
          description: Nackte Domain oder volle URL — wird zur Domain normalisiert.
          example: beispiel.de
        phone:
          type: string
          example: 0571 123456
        email:
          type: string
          format: email
          example: info@beispiel.de
        street:
          type: string
          example: Hauptstr. 1
        city:
          type: string
          example: Espelkamp
        zip:
          type: string
          example: "32339"
        state:
          type: string
          example: NRW
        country:
          type: string
          description: Default „Deutschland", wenn leer.
          example: Deutschland
        industry:
          type: string
          example: Maschinenbau
        company_size:
          type: string
          example: 11-50
        legal_form:
          type: string
          example: GmbH
        register_id:
          type: string
          example: HRB 12345
        description:
          type: string
          example: Freitext / Notizen
        traffic_light_rating:
          type: string
          enum: [green, amber, red]
          description: |
            Optional vorbelegte Webdesign-Ampel: green = heißer Lead (Seite alt /
            Relaunch-Bedarf), amber = unsicher/okay, red = uninteressant (Seite top
            oder Firma inaktiv). Ungültige Werte werden ignoriert (kein Fehler).
            Gesetzte Werte werden als traffic_light_source="api" markiert.
          example: green
        traffic_light_reason:
          type: string
          description: Optionale Begründung zur Ampel.
        contacts:
          type: array
          description: |
            Ansprechpartner. Es werden nur die ersten 3 importiert (weitere werden ignoriert,
            nicht abgelehnt).
          items:
            $ref: '#/components/schemas/ContactInput'
        links:
          type: array
          description: |
            Zusätzliche Webseiten/Profile (Facebook, Instagram, LinkedIn, Google Maps,
            Branchenverzeichnis, …). Werden auch an bereits bestehende (Duplikat-)Leads
            angehängt; idempotent pro (lead_id, url).
          items:
            $ref: '#/components/schemas/LinkInput'

    ContactInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Max Mustermann
        role:
          type: string
          description: Optionale Rolle/Funktion.
          example: Geschäftsführer

    LinkInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Volle URL der Webseite/des Profils.
          example: https://facebook.com/beispiel
        type:
          type: string
          description: |
            Optional. Wird sonst aus der URL erkannt. Bekannte Werte: facebook, instagram,
            linkedin, xing, youtube, tiktok, twitter, google_maps, directory, website, other.
          example: facebook
        label:
          type: string
          description: Optionaler Anzeigename.
          example: Firmen-Facebook

    ImportResult:
      type: object
      required:
        [success, import_log_id, imported, skipped, duplicates, updated, archived, errors, contacts_imported, links_imported, error_details]
      properties:
        success:
          type: boolean
          const: true
        import_log_id:
          type: string
          format: uuid
          description: ID des Import-Batches (für Rückverfolgung/Löschung in der Historie).
        imported:
          type: integer
          description: Neu angelegte Leads.
        skipped:
          type: integer
          description: Übersprungene Zeilen (Duplikate + Pflichtfeld-Fehler).
        duplicates:
          type: integer
          description: Übersprungen, weil bereits vorhanden.
        updated:
          type: integer
          description: Bei dieser API immer 0 (bestehende Leads werden nicht verändert).
        archived:
          type: integer
          description: Bei dieser API immer 0.
        errors:
          type: integer
          description: Abgelehnte Zeilen (z. B. fehlender `company_name`).
        contacts_imported:
          type: integer
        links_imported:
          type: integer
          description: Angelegte Links (neue + an bestehende Leads angehängte).
        error_details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'

    ErrorDetail:
      type: object
      properties:
        row:
          type: integer
          description: 1-basierte Zeilennummer (−1 für Batch-/Insert-Fehler ohne Zeilenbezug).
        field:
          type: string
          example: company_name
        message:
          type: string
          example: Firmenname fehlt

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Lead:
      type: object
      description: |
        Lead-Datensatz. `GET /api/leads` liefert eine kuratierte Feld-Auswahl,
        `GET /api/leads/{id}` den vollständigen Datensatz. Zusätzliche Felder möglich.
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        status: { type: string, example: imported }
        company_name: { type: string, example: Beispiel GmbH }
        website: { type: [string, "null"], example: beispiel.de }
        phone: { type: [string, "null"], example: "+49571123456" }
        email: { type: [string, "null"], example: info@beispiel.de }
        city: { type: [string, "null"], example: Espelkamp }
        vertical: { type: [string, "null"] }
        source_type: { type: [string, "null"], example: manual }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        contacts:
          type: array
          description: Nur bei `include=contacts`.
          items: { $ref: '#/components/schemas/LeadContact' }
        links:
          type: array
          description: Nur bei `include=links`.
          items: { $ref: '#/components/schemas/LeadLink' }

    LeadContact:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        role: { type: [string, "null"] }
        email: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        salutation: { type: [string, "null"] }

    LeadLink:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, example: facebook }
        url: { type: string, example: https://facebook.com/beispiel }
        label: { type: [string, "null"] }

    LeadListResponse:
      type: object
      required: [leads, total, limit, offset]
      properties:
        leads:
          type: array
          items: { $ref: '#/components/schemas/Lead' }
        total:
          type: integer
          description: Gesamtzahl passender Leads (unabhängig von limit/offset).
        limit: { type: integer }
        offset: { type: integer }

    LeadDetailResponse:
      type: object
      required: [lead, contacts, links]
      properties:
        lead: { $ref: '#/components/schemas/Lead' }
        contacts:
          type: array
          items: { $ref: '#/components/schemas/LeadContact' }
        links:
          type: array
          items: { $ref: '#/components/schemas/LeadLink' }

    LeadUpdate:
      type: object
      description: |
        Zu ändernde Stammdaten-Felder (Whitelist). Alle anderen Keys werden ignoriert.
      properties:
        company_name: { type: string }
        website: { type: string }
        phone: { type: string }
        email: { type: string }
        street: { type: string }
        city: { type: string }
        zip: { type: string }
        state: { type: string }
        country: { type: string }
        industry: { type: string }
        company_size: { type: string }
        legal_form: { type: string }
        register_id: { type: string }
        career_page_url: { type: string }
        description: { type: string }
        traffic_light_rating: { type: string, enum: [green, amber, red] }
        traffic_light_score: { type: number }
        traffic_light_source: { type: string }
        traffic_light_rated_at: { type: string, format: date-time }

    UpdateResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true